FIX: Integrated local_ip modified version to the project
This project relies on a modified version of the local_ip crate. The
version has been modified to add functionality to list all network
interfaces and to get the IP address of a specific interface.

The local-ip crate has been integrated as a mod instead of a dependency.

This new version allows to build the project without worrying to
download, compile and integrate to modified local-ip crate.