FIX: Added error management around web server start()
Prior to this commit no check where made to assert if the socket we want
to bind for the web server was free before binding the socket. That
would lead to a panic if the socket was already in use or not available
e.g. port < 1024 with a non root user.

A function fn test_socket(socket: &SocketAddr) -> Result<(), String> as
been added to check if the sockets is free and the fn is called prior to
binding the web server.